+++ apps/web/app/[locale]/(with-layout)/sign-up/brand/utils/signup.ts
